server/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
django-cors-headers==3.11.0
django-nested-admin==3.4.0
djangorestframework==3.13.1
install==1.3.5
Pillow==9.1.0
pycodestyle==2.8.0
python-monkey-business==1.0.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
